WireGuard config loading and basic Win32 implementation. ws/client/desktop/client-desktop#192

Closes #192

See merge request ws/client/desktop/client-desktop!9

+++ backend/windows/windscribe_service/firewallfilter.cpp
@@ -702,7 +702,7 @@ void FirewallFilter::addPermitFilterForAppsIdsExclusiveMode(HANDLE engineHandle,
 	filter.action.type = FWP_ACTION_PERMIT;
 	filter.weight.type = FWP_UINT8;
 	filter.weight.uint8 = 0x02;
-	filter.numFilterConditions = conditions.size();
+	filter.numFilterConditions = static_cast<UINT32>(conditions.size());
 	if (conditions.size() > 0)
 	{
 		filter.filterCondition = &conditions[0];
@@ -777,7 +777,7 @@ void FirewallFilter::addPermitFilterForSplitRoutingWhitelistIps(HANDLE engineHan
 	filter.action.type = FWP_ACTION_PERMIT;
 	filter.weight.type = FWP_UINT8;
 	filter.weight.uint8 = 0x02;
-	filter.numFilterConditions = conditions.size();
+	filter.numFilterConditions = static_cast<UINT32>(conditions.size());
 	if (conditions.size() > 0)
 	{
 		filter.filterCondition = &conditions[0];
